enable strict deabstraction on crashers.swift
